feat: move create invitation request to its own package
