--- backend/app/com/htwhub/ocean/serializers/InstanceSerializer.scala
@@ -1,7 +1,0 @@
-package com.htwhub.ocean.serializers
-
-import com.htwhub.ocean.models.Instance.EngineType
-
-case class CreateInstanceFormData(name: String, engine: EngineType)
-
-object InstanceSerializer {}
