--- backend/app/com/htwhub/ocean/serializers/InvitationSerializer.scala
@@ -1,8 +1,0 @@
-package com.htwhub.ocean.serializers
-
-import com.htwhub.ocean.models.InstanceId
-import com.htwhub.ocean.models.UserId
-
-case class CreateInvitationFormData(instanceId: InstanceId, userId: UserId)
-
-object InvitationSerializer {}
