--- backend/app/com/htwhub/ocean/serializers/RoleSerializer.scala
@@ -1,7 +1,0 @@
-package com.htwhub.ocean.serializers
-
-import com.htwhub.ocean.models.InstanceId
-
-case class CreateRoleFormData(instanceId: InstanceId, roleName: String)
-
-object RoleSerializer {}
